+++ src/services/user.service.ts
@@ -1,5 +1,7 @@
 import IUser from '../interfaces/user';
 import User from '../database/models/Users';
+import NotFound from '../utils/errors/NotFound';
+
 
 export default class UserService {
   constructor(private _UserModel = User) {}
@@ -10,6 +12,12 @@ export default class UserService {
     return user;
   }
 
+  private async getById(id: string) {
+    const user = await this._UserModel.findOne({ where: { id }});
+
+    return user
+  }
+
   public async createUser(user: IUser) {
     const newUser = await this._UserModel.create({ ...user });
 
@@ -17,12 +25,23 @@ export default class UserService {
   }
 
   public async updateEmail(email:string, id:string) {
+    const existUser = await this.getById(id);
+  
+    if(!existUser) {
+      throw new NotFound('pessoa não cadastrada')
+    } 
     const upEmail = await this._UserModel.update({ email }, { where: { id } });
 
     return upEmail;
   }
 
   public async deletUser(id: string) {
+    const existUser = await this.getById(id);
+  
+    if(!existUser) {
+      throw new NotFound('pessoa não cadastrada')
+    } 
+    
     const user = await this._UserModel.destroy({where: { id } });
 
     return user;
